requirements.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
google-auth==2.41.1
google-auth-oauthlib==1.2.2
greenlet==3.2.4
gspread==6.2.1
gspread-formatting==1.2.1
idna==3.11
oauthlib==3.3.1
playwright==1.55.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
